Math::Stats:: Introduce typedef index_type
In #1509, a CI failure was encountered where on Mac ARM64 the shared library would not link, complaining about the absence of __set_fetch_function() for unsigned long datatype. It is believed that this is being caused by the presence of MR::Math::load_matrix() with size_t template type being used for loading pre-defined permutation matrices. To address this, the new typedef MR::Math::Stats::index_type is being introduced to refer to any item being indexed into an array or matrix. A 32-bit unsigned integer should be sufficient for all cases of such (the only possible exception is the calculation of the maximal number of shuffles, for which 64-bit unsigned integers are used, just for the sake of potentially calculating and reporting that number, even if it turns out to be much larger than the number of shuffles actually utilised).

+++ core/math/stats/fwe.cpp
@@ -41,7 +41,7 @@ namespace MR
           for (ssize_t element = 0; element != in.size(); ++element) {
             if (in[element] > 0.0) {
               value_type pvalue = 1.0;
-              for (size_t j = 0; j < size_t(null_dist.size()); ++j) {
+              for (ssize_t j = 0; j < null_dist.size(); ++j) {
                 if (in[element] < null_dist[j]) {
                   pvalue = value_type(j) / value_type(null_dist.size());
                   break;
